refactor to support `join_timeout` and `join_deadline`
Previously one thread would do a potentially long block on
`JoinHandle::join`, but that approach can't support timeouts. Now the
shared thread sets a flag and signals a condvar itself, which lets us
wait with a timeout. This requires injecting exit code into the shared
thread, so we no longer support `SharedThread::new`. The only way to
create a `SharedThread` is `spawn`.